refactor(core-box): migrate shortcut management to global service
The core-box module's shortcut management has been migrated from its internal `ShortcutManager` to the new centralized `GlobalShortcon` service.

This change registers the `core.box.toggle` shortcut (CommandOrControl+E) using `getShortcutService().registerMainShortcut()`. This centralizes shortcut handling, laying the groundwork for future user customization and conflict resolution.

As a result, the local `shortcuts.ts` file and its associated logic have been removed. The module loading order was also adjusted to ensure the shortcut service is available when needed.
